type fix and rename
